@@ -1,12 +1,12 @@
 package br.pedroso.tweetsentiment.app.utils
 
-import android.arch.lifecycle.ViewModel
-import android.arch.lifecycle.ViewModelProvider
-import android.arch.lifecycle.ViewModelProviders
-import android.support.v4.app.FragmentActivity
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModelProviders
+import androidx.fragment.app.FragmentActivity
 
 @Suppress("UNCHECKED_CAST")
-inline fun <reified VM : ViewModel> FragmentActivity.viewModelProvider(
+inline fun <reified VM : ViewModel> androidx.fragment.app.FragmentActivity.viewModelProvider(
         mode: LazyThreadSafetyMode = LazyThreadSafetyMode.NONE,
         crossinline provider: () -> VM) = lazy(mode) {
 
