@@ -1,7 +1,7 @@
 package br.pedroso.tweetsentiment.device.storage.database.room.database
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.RoomDatabase
+import androidx.room.Database
+import androidx.room.RoomDatabase
 import br.pedroso.tweetsentiment.device.storage.database.room.dao.TweetSentimentDao
 import br.pedroso.tweetsentiment.device.storage.database.room.entities.RoomTweet
 import br.pedroso.tweetsentiment.device.storage.database.room.entities.RoomUser
