@@ -1,8 +1,8 @@
 package br.pedroso.tweetsentiment.device.storage.database.room.entities
 
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.ForeignKey
-import android.arch.persistence.room.PrimaryKey
+import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.PrimaryKey
 
 @Entity(tableName = "tweets",
         foreignKeys = [
