@@ -1,7 +1,7 @@
 package br.pedroso.tweetsentiment.device.storage.database.room.entities
 
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.PrimaryKey
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 
 @Entity(tableName = "users")
 data class RoomUser(
