@@ -1,6 +1,6 @@
 package br.pedroso.tweetsentiment.presentation.features.splash
 
-import android.arch.lifecycle.ViewModel
+import androidx.lifecycle.ViewModel
 import br.pedroso.tweetsentiment.domain.entities.ViewState
 import br.pedroso.tweetsentiment.presentation.features.splash.usecases.SplashSyncRegisteredUser
 import br.pedroso.tweetsentiment.presentation.features.splash.viewState.SplashViewStateTransformer
